Refactor: Address Security Vulnerabilities and Improve Configuration
- Verified and confirmed fixes for JWT signature validation, XSS through HTML escaping, and CSRF protection in admin forms, originally identified in codebase audit.
- Hardened Content Security Policy (CSP):
  - Removed 'unsafe-inline' from 'script-src' by externalizing inline JavaScript from admin and homepage templates into static/js/admin.js and static/js/homepage.js respectively.
  - Removed 'unsafe-inline' from 'style-src'. Added a TODO for potential nonce/hash implementation if dynamic theming breaks.
- Improved Configuration Management:
  - ALLOWED_ADMIN_EMAILS and CLOUDFLARE_ACCESS_TEAM_NAME are now sourced from environment variables.
  - Updated wrangler.example.jsonc, .env.example, and config validation logic accordingly.
- Refined Modularity:
  - Confirmed clear separation of concerns in routing.
  - Removed redundant static/index.html.
- Code Quality:
  - Enhanced type safety by replacing 'any' with specific types (CONFIG, DatabaseSubmission, CloudflareAccessUser) in relevant function signatures.
  - Reviewed and confirmed general robustness of error handling.
- Documentation:
  - Updated SETUP.md and docs/admin-access.md to reflect new environment variable configurations.
