+++ .env.example
@@ -1,21 +1,32 @@
 # Environment Variables Template
-# Copy this file to .env and update with your actual values
-# DO NOT commit .env to the repository
+# Copy this file to .env and update with your actual values.
+# DO NOT commit .env to the repository.
+# These variables are typically used for local development with `wrangler dev --env .env`.
+# For production, secrets should be set using `wrangler secret put <NAME>`.
 
-# Cloudflare Configuration
-CLOUDFLARE_ACCOUNT_ID=YOUR_ACCOUNT_ID_HERE
-CLOUDFLARE_DATABASE_ID=YOUR_DATABASE_ID_HERE
+# Cloudflare Configuration (primarily for reference, usually set in wrangler.jsonc or secrets)
+# CLOUDFLARE_ACCOUNT_ID=YOUR_ACCOUNT_ID_HERE
+# CLOUDFLARE_DATABASE_ID=YOUR_DATABASE_ID_HERE
 
-# Email Configuration
-FROM_EMAIL=contact@yourdomain.com
-ADMIN_EMAIL=admin@yourdomain.com
+# Email Configuration (Mailgun)
+# These are also defined in wrangler.jsonc vars, but can be overridden here for local dev if needed.
+# FROM_EMAIL=contact@yourdomain.com
+# ADMIN_EMAIL=admin@yourdomain.com # Recipient for admin notifications
+# MG_DOMAIN=your_mailgun_domain_here # e.g., mg.yourdomain.com
+# MG_API_KEY=your_mailgun_api_key_here # Mailgun Private API Key
 
-# Admin Access (comma-separated list)
-ALLOWED_ADMIN_EMAILS=admin@yourdomain.com,manager@yourdomain.com
+# Admin Access Configuration
+# Comma-separated list of email addresses allowed for basic admin access
+# (if Cloudflare Access is not the sole method or for local dev without Access).
+ALLOWED_ADMIN_EMAILS="admin1@example.com,admin2@example.com"
+
+# Your Cloudflare Access team name (e.g., "your-team" from "your-team.cloudflareaccess.com")
+# Required for JWT signature validation if Cloudflare Access is enabled in the config.
+CLOUDFLARE_ACCESS_TEAM_NAME="your-team-name"
 
 # Application Environment
-ENVIRONMENT=production
+ENVIRONMENT=development # Typically 'development', 'staging', or 'production'
 
-# Optional: API Keys for third-party services
+# Optional: Other API Keys for third-party services
 # SENDGRID_API_KEY=your_sendgrid_api_key_here
-# MAILGUN_API_KEY=your_mailgun_api_key_here
+# OTHER_SERVICE_API_KEY=your_other_service_api_key_here
