+++ src/features/contact/contact-templates.ts
@@ -4,8 +4,9 @@
  */
 
 import { generateThemeCSS } from '../../styles/theme.js';
+import type { CONFIG } from '../../config/index.js'; // Import CONFIG
 
-export function getContactFormHTML(config: any): string {
+export function getContactFormHTML(config: CONFIG): string {
 	return `<!DOCTYPE html>
 <html lang="en">
 <head>
