+++ src/features/website/website-templates.ts
@@ -2,8 +2,9 @@
  * Website Templates
  * Enhanced brand presence templates
  */
+import type { CONFIG } from '../../config/index.js'; // Import CONFIG
 
-export async function getHomepageHTML(corsHeaders: Record<string, string>, config: any): Promise<Response> {
+export async function getHomepageHTML(corsHeaders: Record<string, string>, config: CONFIG): Promise<Response> {
 	// Enhanced Atlas Divisions homepage with improved brand presence
 	const html = `<!DOCTYPE html>
 <html lang="en">
@@ -16,30 +17,6 @@ export async function getHomepageHTML(corsHeaders: Record<string, string>, confi
     
     <!-- Google Fonts -->
     <link href="https://fonts.googleapis.com/css2?family=Inter:wght@300;400;500;600;700&family=Montserrat:wght@400;500;600;700;800&display=swap" rel="stylesheet">
-    
-    <!-- Three.js ES6 Module from CDNjs -->
-    <script type="module">
-        import * as THREE from 'https://cdnjs.cloudflare.com/ajax/libs/three.js/0.177.0/three.module.min.js';
-        window.THREE = THREE;
-        
-        // Set up Atlas Globe initialization function with enhanced error handling
-        window.initAtlasGlobeWhenReady = function() {
-            if (typeof AtlasGlobe !== 'undefined' && typeof THREE !== 'undefined') {
-                try {
-                    window.atlasGlobe = new AtlasGlobe();
-                    console.log('Atlas Globe initialized successfully');
-                } catch (error) {
-                    console.error('Atlas Globe initialization failed:', error);
-                    // Fallback to static globe emoji
-                    const container = document.getElementById('globe-container');
-                    if (container) {
-                        container.innerHTML = '<div style="width:100%;height:100%;display:flex;align-items:center;justify-content:center;color:#d4af37;font-size:8rem;animation:float 6s ease-in-out infinite;">🌍</div>';
-                    }
-                }
-            }
-        };
-        
-    </script>
     
     <style>
         :root {
@@ -734,442 +711,7 @@ export async function getHomepageHTML(corsHeaders: Record<string, string>, confi
         </div>
     </section>
 
-    <script>
-        // Atlas Divisions Globe Class - Enhanced Animation Implementation
-        class AtlasGlobe {
-            constructor() {
-                this.scene = null;
-                this.camera = null;
-                this.renderer = null;
-                this.globeMesh = null;
-                
-                this.isDragging = false;
-                this.previousMousePosition = { x: 0, y: 0 };
-                this.rotationVelocity = { x: 0, y: 0 };
-                this.autoRotationSpeed = 0.005; // Fixed consistent rate for west-to-east rotation
-                this.friction = 0.95; // Slightly more friction for smoother transitions
-                this.rotationSpeed = 0.008; // Adjusted for left-right only interaction
-                
-                // Atlas Divisions styling
-                this.atlasColors = {
-                    ocean: '#001122',
-                    land: '#d4af37',   // Atlas gold
-                    stroke: '#cd7f32', // Atlas bronze
-                    atmosphere: 0xd4af37, // Gold atmosphere
-                    light: 0xd4af37    // Gold lighting
-                };
-                
-                this.init();
-            }
-            
-            init() {
-                if (typeof THREE === 'undefined') {
-                    console.error('Three.js not loaded');
-                    const container = document.getElementById('globe-container');
-                    if (container) {
-                        container.innerHTML = '<div style="width:100%;height:100%;display:flex;align-items:center;justify-content:center;color:#d4af37;font-size:4rem;">🌍</div>';
-                    }
-                    return;
-                }
-                
-                this.createScene();
-                this.setupEventListeners();
-            }
-            
-            createScene() {
-                const container = document.getElementById('globe-container');
-                if (!container) return;
-                
-                // Force container to update its computed styles
-                container.style.display = container.style.display;
-                
-                // Wait a moment for CSS to apply, then get dimensions
-                setTimeout(() => {
-                    const width = container.offsetWidth || container.clientWidth;
-                    const height = container.offsetHeight || container.clientHeight;
-                    
-                    console.log('Globe container dimensions:', width, 'x', height);
-                    
-                    this.scene = new THREE.Scene();
-                    this.camera = new THREE.PerspectiveCamera(75, width / height, 0.1, 1000);
-                    this.renderer = new THREE.WebGLRenderer({ antialias: true, alpha: true });
-                    
-                    this.renderer.setSize(width, height);
-                    this.renderer.setClearColor(0x000000, 0);
-                    container.appendChild(this.renderer.domElement);
-                    
-                    this.camera.position.z = 4;
-                    
-                    // Create globe after scene is ready
-                    this.createGlobe();
-                    this.setupLighting();
-                    this.animate();
-                }, 100);
-            }
-            
-            async createGlobe() {
-                const geometry = new THREE.SphereGeometry(1.5, 64, 64);
-                const texture = await this.createAtlasTexture();
-                
-                const material = new THREE.MeshPhongMaterial({
-                    map: texture,
-                    transparent: true,
-                    opacity: 0.9
-                });
-                
-                this.globeMesh = new THREE.Mesh(geometry, material);
-                this.globeMesh.rotation.x = 0.1;
-                this.scene.add(this.globeMesh);
-                
-                this.createAtmosphere();
-                
-                // Start with consistent west-to-east auto-rotation (Y-axis only)
-                this.rotationVelocity.x = 0; // No vertical rotation
-                this.rotationVelocity.y = this.autoRotationSpeed;
-            }
-            
-            createAtmosphere() {
-                const atmosphereGeometry = new THREE.SphereGeometry(1.6, 64, 64);
-                const atmosphereMaterial = new THREE.MeshBasicMaterial({
-                    color: this.atlasColors.atmosphere,
-                    transparent: true,
-                    opacity: 0.1,
-                    side: THREE.BackSide
-                });
-                const atmosphere = new THREE.Mesh(atmosphereGeometry, atmosphereMaterial);
-                this.scene.add(atmosphere);
-            }
-            
-            setupLighting() {
-                const ambientLight = new THREE.AmbientLight(this.atlasColors.light, 0.4);
-                this.scene.add(ambientLight);
-                
-                const directionalLight = new THREE.DirectionalLight(this.atlasColors.light, 0.8);
-                directionalLight.position.set(5, 3, 5);
-                this.scene.add(directionalLight);
-            }
-            
-            async createAtlasTexture() {
-                const canvas = document.createElement('canvas');
-                canvas.width = 2048;
-                canvas.height = 1024;
-                const ctx = canvas.getContext('2d');
-                
-                // Ocean background
-                ctx.fillStyle = this.atlasColors.ocean;
-                ctx.fillRect(0, 0, canvas.width, canvas.height);
-                
-                try {
-                    const response = await fetch('https://raw.githubusercontent.com/holtzy/D3-graph-gallery/master/DATA/world.geojson');
-                    const geoData = await response.json();
-                    this.drawWorldMap(ctx, geoData, canvas.width, canvas.height);
-                } catch (error) {
-                    console.log('Using fallback map');
-                    this.drawFallbackMap(ctx, canvas.width, canvas.height);
-                }
-                
-                const texture = new THREE.CanvasTexture(canvas);
-                texture.wrapS = THREE.RepeatWrapping;
-                texture.wrapT = THREE.ClampToEdgeWrapping;
-                return texture;
-            }
-            
-            drawWorldMap(ctx, geoData, width, height) {
-                ctx.fillStyle = this.atlasColors.land;
-                ctx.strokeStyle = this.atlasColors.stroke;
-                ctx.lineWidth = 1;
-                
-                geoData.features.forEach(feature => {
-                    if (feature.geometry.type === 'Polygon') {
-                        this.drawPolygon(ctx, feature.geometry.coordinates, width, height);
-                    } else if (feature.geometry.type === 'MultiPolygon') {
-                        feature.geometry.coordinates.forEach(polygon => {
-                            this.drawPolygon(ctx, polygon, width, height);
-                        });
-                    }
-                });
-            }
-            
-            drawPolygon(ctx, coordinates, width, height) {
-                coordinates.forEach(ring => {
-                    ctx.beginPath();
-                    ring.forEach((coord, index) => {
-                        const x = ((coord[0] + 180) / 360) * width;
-                        const y = ((90 - coord[1]) / 180) * height;
-                        
-                        if (index === 0) {
-                            ctx.moveTo(x, y);
-                        } else {
-                            ctx.lineTo(x, y);
-                        }
-                    });
-                    ctx.closePath();
-                    ctx.fill();
-                    ctx.stroke();
-                });
-            }
-            
-            drawFallbackMap(ctx, width, height) {
-                const continents = [
-                    { x: 0.2, y: 0.3, w: 0.25, h: 0.4 },
-                    { x: 0.25, y: 0.5, w: 0.15, h: 0.35 },
-                    { x: 0.48, y: 0.25, w: 0.12, h: 0.15 },
-                    { x: 0.5, y: 0.35, w: 0.15, h: 0.4 },
-                    { x: 0.6, y: 0.2, w: 0.3, h: 0.35 },
-                    { x: 0.75, y: 0.65, w: 0.12, h: 0.1 }
-                ];
-                
-                ctx.fillStyle = this.atlasColors.land;
-                continents.forEach(continent => {
-                    ctx.fillRect(
-                        continent.x * width,
-                        continent.y * height,
-                        continent.w * width,
-                        continent.h * height
-                    );
-                });
-            }
-            
-            setupEventListeners() {
-                const container = document.getElementById('globe-container');
-                if (!container) return;
-                
-                container.style.cursor = 'grab';
-                
-                container.addEventListener('mousedown', this.onMouseDown.bind(this));
-                container.addEventListener('mousemove', this.onMouseMove.bind(this));
-                container.addEventListener('mouseup', this.onMouseUp.bind(this));
-                container.addEventListener('mouseleave', this.onMouseLeave.bind(this));
-                
-                container.addEventListener('touchstart', this.onTouchStart.bind(this));
-                container.addEventListener('touchmove', this.onTouchMove.bind(this));
-                container.addEventListener('touchend', this.onTouchEnd.bind(this));
-                
-                window.addEventListener('resize', this.onWindowResize.bind(this));
-                
-                // Add ResizeObserver to watch container size changes
-                if (window.ResizeObserver) {
-                    const resizeObserver = new ResizeObserver((entries) => {
-                        for (const entry of entries) {
-                            if (entry.target.id === 'globe-container') {
-                                this.onWindowResize();
-                            }
-                        }
-                    });
-                    resizeObserver.observe(container);
-                }
-            }
-            
-            onMouseDown(event) {
-                this.startDragging(event.clientX, event.clientY);
-                document.getElementById('globe-container').style.cursor = 'grabbing';
-            }
-            
-            onMouseMove(event) {
-                if (this.isDragging) {
-                    this.updateRotation(event.clientX, event.clientY);
-                }
-            }
-            
-            onMouseUp() {
-                this.stopDragging();
-                document.getElementById('globe-container').style.cursor = 'grab';
-            }
-            
-            onMouseLeave() {
-                this.stopDragging();
-                document.getElementById('globe-container').style.cursor = 'grab';
-            }
-            
-            onTouchStart(event) {
-                event.preventDefault();
-                if (event.touches.length === 1) {
-                    const touch = event.touches[0];
-                    this.startDragging(touch.clientX, touch.clientY);
-                }
-            }
-            
-            onTouchMove(event) {
-                event.preventDefault();
-                if (this.isDragging && event.touches.length === 1) {
-                    const touch = event.touches[0];
-                    this.updateRotation(touch.clientX, touch.clientY);
-                }
-            }
-            
-            onTouchEnd(event) {
-                event.preventDefault();
-                this.stopDragging();
-            }
-            
-            startDragging(clientX, clientY) {
-                this.isDragging = true;
-                const container = document.getElementById('globe-container');
-                const rect = container.getBoundingClientRect();
-                this.previousMousePosition = {
-                    x: clientX - rect.left,
-                    y: clientY - rect.top
-                };
-            }
-            
-            updateRotation(clientX, clientY) {
-                const container = document.getElementById('globe-container');
-                const rect = container.getBoundingClientRect();
-                const currentMousePosition = {
-                    x: clientX - rect.left,
-                    y: clientY - rect.top
-                };
-                
-                const deltaX = currentMousePosition.x - this.previousMousePosition.x;
-                // const deltaY = currentMousePosition.y - this.previousMousePosition.y; // Ignore Y movement
-                
-                // Only allow left-right (horizontal) rotation - Y-axis rotation only
-                this.rotationVelocity.x = 0; // Disable vertical rotation completely
-                this.rotationVelocity.y = deltaX * this.rotationSpeed; // Only horizontal rotation
-                
-                this.previousMousePosition = currentMousePosition;
-            }
-            
-            stopDragging() {
-                this.isDragging = false;
-            }
-            
-            onWindowResize() {
-                const container = document.getElementById('globe-container');
-                if (!container || !this.renderer || !this.camera) return;
-                
-                const width = container.offsetWidth || container.clientWidth;
-                const height = container.offsetHeight || container.clientHeight;
-                
-                console.log('Resizing globe canvas to:', width, 'x', height);
-                
-                this.camera.aspect = width / height;
-                this.camera.updateProjectionMatrix();
-                this.renderer.setSize(width, height);
-            }
-            
-            animate() {
-                requestAnimationFrame(this.animate.bind(this));
-                
-                if (this.globeMesh) {
-                    if (this.isDragging) {
-                        // When dragging, only apply Y-axis rotation from user interaction
-                        this.globeMesh.rotation.y += this.rotationVelocity.y;
-                        // Keep X rotation fixed (no up-down rotation)
-                    } else {
-                        // When not dragging, apply friction to user velocity
-                        this.rotationVelocity.y *= this.friction;
-                        
-                        // Apply user velocity if still significant
-                        if (Math.abs(this.rotationVelocity.y) > 0.001) {
-                            this.globeMesh.rotation.y += this.rotationVelocity.y;
-                        } else {
-                            // Return to consistent fixed auto-rotation (west to east)
-                            this.rotationVelocity.y = 0;
-                            this.globeMesh.rotation.y += this.autoRotationSpeed;
-                        }
-                    }
-                    
-                    // Keep X rotation fixed - no vertical rotation allowed
-                    this.globeMesh.rotation.x = 0.1; // Slight tilt for better viewing angle
-                }
-                
-                this.renderer.render(this.scene, this.camera);
-            }
-        }
-        
-        // Initialize globe
-        let atlasGlobe = null;
-        
-        // Copy email functionality
-        function copyEmail() {
-            const email = '${config.company.primaryEmail}';
-            
-            if (navigator.clipboard && navigator.clipboard.writeText) {
-                navigator.clipboard.writeText(email).then(() => {
-                    showCopyFeedback();
-                }).catch(() => {
-                    fallbackCopyMethod(email);
-                });
-            } else {
-                fallbackCopyMethod(email);
-            }
-        }
-        
-        function fallbackCopyMethod(email) {
-            const textArea = document.createElement('textarea');
-            textArea.value = email;
-            textArea.style.position = 'fixed';
-            textArea.style.opacity = '0';
-            document.body.appendChild(textArea);
-            textArea.select();
-            
-            try {
-                document.execCommand('copy');
-                showCopyFeedback();
-            } catch (err) {
-                console.error('Failed to copy email:', err);
-            }
-            
-            document.body.removeChild(textArea);
-        }
-        
-        function showCopyFeedback() {
-            const feedback = document.getElementById('copy-feedback');
-            if (feedback) {
-                feedback.style.display = 'inline';
-                setTimeout(() => {
-                    feedback.style.display = 'none';
-                }, 2000);
-            }
-        }
-        
-        // Smooth scrolling for navigation links
-        document.querySelectorAll('a[href^="#"]').forEach(anchor => {
-            anchor.addEventListener('click', function (e) {
-                e.preventDefault();
-                const target = document.querySelector(this.getAttribute('href'));
-                if (target) {
-                    target.scrollIntoView({
-                        behavior: 'smooth',
-                        block: 'start'
-                    });
-                }
-            });
-        });
-        
-        // Initialize Atlas Globe when page loads with progressive enhancement
-        window.addEventListener('load', function() {
-            // Initialize the Atlas Globe with retry logic (no loading message that breaks formatting)
-            function initializeGlobe(retries = 3) {
-                if (typeof window.initAtlasGlobeWhenReady === 'function') {
-                    window.initAtlasGlobeWhenReady();
-                } else if (retries > 0) {
-                    setTimeout(() => initializeGlobe(retries - 1), 200);
-                } else {
-                    // Final fallback after all retries - just show static globe emoji
-                    const container = document.getElementById('globe-container');
-                    if (container) {
-                        container.innerHTML = '<div style="width:100%;height:100%;display:flex;align-items:center;justify-content:center;color:#d4af37;font-size:8rem;animation:float 6s ease-in-out infinite;">🌍</div>';
-                    }
-                }
-            }
-            
-            initializeGlobe();
-        });
-        
-        // Cleanup on page unload
-        window.addEventListener('beforeunload', () => {
-            if (atlasGlobe && atlasGlobe.renderer) {
-                // Clean up Three.js resources
-                const container = document.getElementById('globe-container');
-                if (container && atlasGlobe.renderer.domElement) {
-                    container.removeChild(atlasGlobe.renderer.domElement);
-                }
-                atlasGlobe.renderer.dispose();
-            }
-        });
-    </script>
+    <script type="module" src="/js/homepage.js" defer></script>
 </body>
 </html>`;
